weather-extractor/pyproject.toml
[tool.poetry]
name = "weather-extractor"
version = "1.0.0"
description = "A simple weather data extractor"
authors = ["Mathias Lohne <mathias.lohne@cognite.com>"]
packages = [
    { include="met_extractor" },
]

[tool.poetry.dependencies]
python = "^3.8"
cognite-extractor-utils = "^2.3.0-beta1"
cognite-sdk-experimental = "^0.80.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
weather_extractor = "met_extractor.__main__:main"
